+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `queueMaxSize` option to tasks queue options.
+
 ## [2.6.31] - 2023-08-20
 
 ### Fixed